Split airsim offboard and mavros offboard

+++ control/launch/launch-offb.sh
@@ -1,8 +1,26 @@
 #!/bin/bash
 
-killall gzclient
-killall gzserver
+if [ "$#" -ne 1 ]; then
+    echo "Please input argument \"airsim\" or \"mavros\""
+    exit 0
+fi
 
-source ./launch-common.sh
+if [ "$1" == "airsim" ]; then
+    input="airsim"
+elif [ "$1" == "mavros" ]; then
+    input="mavros"
+else
+    echo "Please input argument \"airsim\" or \"mavros\""
+    exit 0
+fi
 
-roslaunch control mavros_offb_node.launch
+if [ $input == "airsim" ]; then
+    roslaunch control control.launch do_airsim:="true"
+elif [ $input == "mavros" ]; then
+    killall gzclient
+    killall gzserver
+
+    source ./launch-common.sh
+
+    roslaunch control control.launch do_airsim:="false"
+fi
